refactor: arc and mutex eliminations
